+++ changelog.md
@@ -1,8 +1,12 @@
 3.5.14.0 (unreleased)
 ========
 - @parsonsmatt
+    - [#415](https://github.com/bitemyapp/esqueleto/pull/415)
+        - Export the `SqlSelect` type from `Database.Esqueleto.Experimental`
     - [#414](https://github.com/bitemyapp/esqueleto/pull/414)
         - Derive `Foldable` and `Traversable` for `Value`.
+    - [#416](https://github.com/bitemyapp/esqueleto/pull/416)
+        - Derive `Functor` and `Bifunctor` for `:&`
 - @matthewbauer
     - [#341](https://github.com/bitemyapp/esqueleto/pull/341/)
         - Add functions for `NULLS FIRST` and `NULLS LAST` in the Postgresql